[Bronze V] Title: Multiply, Time: 112 ms, Memory: 14136 KB, Score: 40 point -BaekjoonHub

+++ 백준/Bronze/22193. Multiply/Multiply.java
@@ -6,6 +6,6 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         br.readLine();
-        System.out.println(Integer.parseInt(br.readLine()) * Integer.parseInt(br.readLine()));
+        System.out.println(Long.parseLong(br.readLine()) * Long.parseLong(br.readLine()));
     }
 }
